10-backend-apis/02-django/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
